Use system default for css dark/light theme choice
Rather than defaulting to the dark theme, let's default to the user's
system setting so that the light or dark theme is selected
automatically.

If there is no cookie present for the theme selection default to the
system theme.

If there is a cookie set use that regardless of system theme.

Regarding the code, we are using reacts `dangerouslySetInnerHTML`
property https://react.dev/reference/react-dom/components/common#dangerously-setting-the-inner-html
but we are using it in a safe way. It only interpolates `theme` which is
`'light'`, `'dark'`, or `null` - values we control from the cookie. Even
if someone tampered with the cookie, the only values that matter are
`'light'` or `'dark'`. Anything else just defaults to dark mode. So its
safe to use in this case.

Resolves #172

Depends on #177